requirements.txt
ephem
satellitetle>=0.5.0
requests
lxml
tqdm
python-decouple
